fix(ExpressionContext): change startSeconds type from Integer to Double

+++ src/main/java/com/example/auth/expression/dto/ContextDto.java
@@ -10,6 +10,6 @@ public record ContextDto(
     UUID video_uuid,
     String videoTitle,
     List<String> matchedForms,
-    Integer startSeconds
+    Double startSeconds
 ) {
 }
